+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/notifications/NotificationServiceTest.scala
@@ -7,7 +7,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatestplus.mockito.MockitoSugar
 import pl.touk.nussknacker.engine.api.ProcessVersion
-import pl.touk.nussknacker.engine.api.deployment.{DeployedScenarioData, DeploymentManager}
+import pl.touk.nussknacker.engine.api.deployment.{DataFreshnessPolicy, DeployedScenarioData, DeploymentManager, ProcessState, WithDataFreshnessStatus}
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
@@ -83,7 +83,8 @@ class NotificationServiceTest extends AnyFunSuite with Matchers with PatientScal
   }
 
   private def createServices(deploymentManager: DeploymentManager, processName: ProcessName) = {
-    when(deploymentManager.findJobStatus(any[ProcessName])).thenReturn(Future.successful(None))
+    when(deploymentManager.getProcessState(any[ProcessName])(any[DataFreshnessPolicy]))
+      .thenReturn(Future.successful(WithDataFreshnessStatus(Option.empty[ProcessState], cached = false)))
     val managerDispatcher = mock[DeploymentManagerDispatcher]
     when(managerDispatcher.deploymentManager(any[String])).thenReturn(deploymentManager)
     val config = NotificationConfig(20 minutes)