+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManagerTest.scala
@@ -8,7 +8,7 @@ import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.api.{MetaData, ProcessVersion, StreamMetaData}
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
-import pl.touk.nussknacker.engine.api.deployment.ProcessActionType
+import pl.touk.nussknacker.engine.api.deployment.{DataFreshnessPolicy, ProcessActionType}
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.{DeploymentData, User}
@@ -17,6 +17,7 @@ import pl.touk.nussknacker.engine.management.periodic.PeriodicStateStatus.{Sched
 import pl.touk.nussknacker.engine.management.periodic.model.PeriodicProcessDeploymentStatus
 import pl.touk.nussknacker.engine.management.periodic.service.{DefaultAdditionalDeploymentDataProvider, EmptyListener, ProcessConfigEnricher}
 import pl.touk.nussknacker.test.PatientScalaFutures
+
 import java.time.Clock
 
 class PeriodicDeploymentManagerTest extends AnyFunSuite
@@ -26,6 +27,8 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
   with Inside
   with TableDrivenPropertyChecks
   with PatientScalaFutures {
+
+  protected implicit val freshnessPolicy: DataFreshnessPolicy = DataFreshnessPolicy.Fresh
 
   import org.scalatest.LoneElement._
 
@@ -57,31 +60,31 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
       toClose = () => ()
     )
 
-    def getAllowedActions: List[ProcessActionType] = periodicDeploymentManager.findJobStatus(processName).futureValue.value.allowedActions
+    def getAllowedActions: List[ProcessActionType] = periodicDeploymentManager.getProcessState(processName).futureValue.value.value.allowedActions
   }
 
-  test("findJobStatus - should return none for no job") {
+  test("getProcessState - should return none for no job") {
     val f = new Fixture
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     state shouldBe Symbol("empty")
   }
 
-  test("findJobStatus - should return none for scenario with different processing type") {
+  test("getProcessState - should return none for scenario with different processing type") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Scheduled, processingType = "other")
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     state shouldBe Symbol("empty")
   }
 
-  test("findJobStatus - should be scheduled when scenario scheduled and no job on Flink") {
+  test("getProcessState - should be scheduled when scenario scheduled and no job on Flink") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Scheduled)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe a[ScheduledStatus]
@@ -89,47 +92,47 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     state.value.allowedActions shouldBe List(ProcessActionType.Cancel, ProcessActionType.Deploy)
   }
 
-  test("findJobStatus - should be scheduled when scenario scheduled and job finished on Flink") {
+  test("getProcessState - should be scheduled when scenario scheduled and job finished on Flink") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Scheduled)
     f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Finished)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe a[ScheduledStatus]
     state.value.allowedActions shouldBe List(ProcessActionType.Cancel, ProcessActionType.Deploy)
   }
 
-  test("findJobStatus - should be running when scenario deployed and job running on Flink") {
+  test("getProcessState - should be running when scenario deployed and job running on Flink") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed)
     f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Running)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe FlinkStateStatus.Running
     state.value.allowedActions shouldBe List(ProcessActionType.Cancel)
   }
 
-  test("findJobStatus - should be waiting for reschedule if job finished on Flink but scenario is still deployed") {
+  test("getProcessState - should be waiting for reschedule if job finished on Flink but scenario is still deployed") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed)
     f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Finished)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe WaitingForScheduleStatus
     state.value.allowedActions shouldBe List(ProcessActionType.Cancel)
   }
 
-  test("findJobStatus - should be failed after unsuccessful deployment") {
+  test("getProcessState - should be failed after unsuccessful deployment") {
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Failed)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe SimpleStateStatus.Failed
@@ -185,7 +188,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed)
     f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Failed)
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value
 
     val status = state.value.status
     status shouldBe SimpleStateStatus.Failed
@@ -196,7 +199,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     val f = new Fixture
     f.repository.addActiveProcess(processName, PeriodicProcessDeploymentStatus.Deployed)
     f.delegateDeploymentManagerStub.setStateStatus(FlinkStateStatus.Failed)
-    val failedProcessState = f.periodicDeploymentManager.findJobStatus(processName).futureValue.value
+    val failedProcessState = f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value
     failedProcessState.status shouldBe FlinkStateStatus.Failed
     failedProcessState.allowedActions shouldBe List(ProcessActionType.Cancel) // redeploy is blocked in GUI but API allows it
 
@@ -204,7 +207,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
 
     f.repository.processEntities.map(_.active) shouldBe List(false, true)
     f.repository.deploymentEntities.map(_.status) shouldBe List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Scheduled)
-    val scheduledProcessState = f.periodicDeploymentManager.findJobStatus(processName).futureValue.value
+    val scheduledProcessState = f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value
     // Previous job is still visible as Failed.
     scheduledProcessState.status shouldBe a[ScheduledStatus]
     scheduledProcessState.allowedActions shouldBe List(ProcessActionType.Cancel, ProcessActionType.Deploy)
@@ -253,7 +256,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     //this one is cyclically called by RescheduleActor
     f.periodicProcessService.handleFinished.futureValue
 
-    f.periodicDeploymentManager.findJobStatus(processName).futureValue.get.status shouldBe SimpleStateStatus.Failed
+    f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value.status shouldBe SimpleStateStatus.Failed
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
     f.repository.processEntities.loneElement.active shouldBe true
 
@@ -261,7 +264,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
-    f.periodicDeploymentManager.findJobStatus(processName).futureValue.get.status shouldBe SimpleStateStatus.Canceled
+    f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value.status shouldBe SimpleStateStatus.Canceled
   }
 
   test("should reschedule failed job after RescheduleActor handles finished when configured") {
@@ -272,7 +275,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     //this one is cyclically called by RescheduleActor
     f.periodicProcessService.handleFinished.futureValue
 
-    val state = f.periodicDeploymentManager.findJobStatus(processName).futureValue.get
+    val state = f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value
     state.status shouldBe a[ScheduledStatus]
     f.repository.deploymentEntities.map(_.status) shouldBe List(PeriodicProcessDeploymentStatus.Failed, PeriodicProcessDeploymentStatus.Scheduled)
     f.repository.processEntities.loneElement.active shouldBe true
@@ -287,7 +290,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
-    f.periodicDeploymentManager.findJobStatus(processName).futureValue.get.status shouldBe SimpleStateStatus.Canceled
+    f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value.status shouldBe SimpleStateStatus.Canceled
   }
 
   test("should cancel failed scenario after disappeared from Flink console") {
@@ -301,7 +304,7 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
     //after some time Flink stops returning job status
     f.delegateDeploymentManagerStub.jobStatus = None
 
-    f.periodicDeploymentManager.findJobStatus(processName).futureValue.get.status shouldBe SimpleStateStatus.Failed
+    f.periodicDeploymentManager.getProcessState(processName).futureValue.value.value.status shouldBe SimpleStateStatus.Failed
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
     f.repository.processEntities.loneElement.active shouldBe true
 
@@ -309,6 +312,6 @@ class PeriodicDeploymentManagerTest extends AnyFunSuite
 
     f.repository.processEntities.loneElement.active shouldBe false
     f.repository.deploymentEntities.loneElement.status shouldBe PeriodicProcessDeploymentStatus.Failed
-    f.periodicDeploymentManager.findJobStatus(processName).futureValue shouldBe None
+    f.periodicDeploymentManager.getProcessState(processName).futureValue.value shouldBe None
   }
 }